🔐 Added module in the JRE to support SSL connection

+++ src/main/kotlin/io/vinicius/rmd/Main.kt
@@ -5,46 +5,59 @@ import io.vinicius.rmd.model.Response
 import io.vinicius.rmd.model.Submission
 import io.vinicius.rmd.util.Fetch
 import io.vinicius.rmd.util.Shell
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.runBlocking
 import java.io.File
 import java.time.LocalDateTime
 import java.time.ZoneOffset
 import java.time.format.DateTimeFormatter
 import kotlin.math.ceil
+import kotlin.time.Duration.Companion.seconds
 
 fun main(args: Array<String>) {
-    val user = System.getenv("RMD_USER")
-    val limit = System.getenv("RMD_LIMIT").toInt()
+    val user: String? = System.getenv("RMD_USERR")
+    val limit: Int? = System.getenv("RMD_LIMIT")?.toInt()
 
-    println("⬇ Downloading $limit media from user [$user]...\n")
+    if (user == null || limit == null) {
+        error("🧨".em + " Missing the environment variable RMD_USER or RMD_LIMIT")
+    }
 
     val submissions = getSubmissions(user, limit)
     val downloads = downloadMedia(user, submissions)
 
-    cleanFailedDownloads(user, downloads)
+    removeDuplicates(user, downloads)
     println("\n🌟 Done!")
 }
 
 private fun getSubmissions(user: String, limit: Int): Set<Submission> {
     val fetch = Fetch()
-    val submissions = mutableListOf<Submission>()
+    val submissions = mutableSetOf<Submission>()
     var before = LocalDateTime.now().toEpochSecond(ZoneOffset.UTC)
 
-    repeat(ceil(limit / 100f).toInt()) {
+    print("📝".em + " Collecting $limit posts from user [$user] ")
+
+    repeat(ceil(limit / 250f).toInt()) {
         val url = createUrl(user, before)
         val list = fetch.get<Response>(url).data
         submissions.addAll(list)
 
         // Getting the last "before"
         before = list.last().created
+
+        runBlocking {
+            print(".")
+            delay(2.seconds)
+        }
     }
 
+    println(" $submissions/$limit unique posts found\n")
     return submissions.toSet()
 }
 
 private fun createUrl(user: String, before: Long): String {
     val baseUrl = "https://api.pushshift.io/reddit/submission/search"
     val fields = "author,created_utc,domain,post_hint,url"
-    return "${baseUrl}?author=${user}&fields=${fields}&before=${before}&size=100"
+    return "${baseUrl}?author=${user}&fields=${fields}&before=${before}"
 }
 
 private fun downloadMedia(user: String, submissions: Set<Submission>): List<Download> {
@@ -59,11 +72,11 @@ private fun downloadMedia(user: String, submissions: Set<Submission>): List<Down
 
         val success = if (submission.postHint == "image") {
             fileName = "$baseFile-$number.jpg"
-            print("🖼 [$number] Downloading image ${submission.url} ".padEnd(100, '.'))
+            print("🖼".em + " [$number] Downloading image ${submission.url} ".padEnd(100, '.'))
             shell.downloadImage(submission.url, fileName)
         } else {
             fileName = "$baseFile-$number.mp4"
-            print("🎥 [$number] Downloading video ${submission.url} ".padEnd(100, '.'))
+            print("🎥".em + " [$number] Downloading video ${submission.url} ".padEnd(100, '.'))
             shell.downloadVideo(submission.url, fileName)
         }
 
@@ -71,22 +84,19 @@ private fun downloadMedia(user: String, submissions: Set<Submission>): List<Down
         val hash = shell.calculateHash(fileName).orEmpty()
         downloads.add(Download(fileName, success, hash))
 
-        println(if (success) " ✅ [Success]" else " ❌ [Error]")
+        println(if (success) " ✅ [Success]" else " ❌ [Failure]")
     }
 
     return downloads
 }
 
-private fun cleanFailedDownloads(user: String, downloads: List<Download>) {
-    println("\n🗑 Removing duplicated and failed downloads...")
+private fun removeDuplicates(user: String, downloads: List<Download>) {
+    println("\n" + "🗑".em + " Removing duplicated downloads...")
 
     // Removing 0-byte files
     downloads.forEach {
         val file = File("/tmp/rmd/$user", it.fileName)
-        if (file.exists() && file.length() == 0L) {
-            println("ZER: ${file.absoluteFile}")
-            file.delete()
-        }
+        if (file.exists() && file.length() == 0L) file.delete()
     }
 
     // Removing duplicates
@@ -94,7 +104,7 @@ private fun cleanFailedDownloads(user: String, downloads: List<Download>) {
         it.drop(1).forEach { download ->
             val file = File("/tmp/rmd/$user", download.fileName)
             if (file.exists()) {
-                println("DUP: ${file.absoluteFile}")
+                println(file.absoluteFile)
                 file.delete()
             }
         }
